driver to create sim mesh
